# backend/src/main/java/ru/urasha/callmeani/dream_marketplace/dto/AuthResponse.java
package ru.urasha.callmeani.dream_marketplace.dto;

public record AuthResponse(String token, String tokenType, UserDto user) {
    public static AuthResponse bearer(String token, UserDto user) {
        return new AuthResponse(token, "Bearer", user);
    }
}